Added more info
