+++ Chess.View/StatusAndMode/StatusModeListViewView.xaml
@@ -101,7 +101,7 @@
                     <RowDefinition x:Name="HorizontalSplitterRow" Height="5" />
                     <RowDefinition x:Name="ChessMovesNotesRow" Height="{Binding ChessMovesNotesRowHeight}" />
                 </Grid.RowDefinitions>
-                <Border CornerRadius="5" Background="LightGray" Margin="10,5,10,10" x:Name="ChessMovesListViewBorder" Focusable="True" >
+                <Border Grid.Row="0" CornerRadius="5" Background="LightGray" Margin="10,5,10,10" x:Name="ChessMovesListViewBorder" Focusable="True" >
                     <ListView x:Name="myListView" Grid.Row="0"
                         ItemsSource="{Binding Path=Board.ChessMoveSequence.ChessMoves}"
                         Height="Auto" Margin="10 5 10 10 " Width="Auto"
