+++ Chess.View/Window/MainWindow.xaml.cs
@@ -56,20 +56,6 @@ namespace Chess.View.Window
 
             if (!string.IsNullOrWhiteSpace(ChessAppSettings.Default.ConsoleColumnWidth))
                 ConsoleColumn.Width = (GridLength)gridLengthConverter.ConvertFromString(ChessAppSettings.Default.ConsoleColumnWidth);
-
-
-
-            //if (!string.IsNullOrWhiteSpace(ChessAppSettings.Default.ChessMovesListViewRowHeight))
-            //    ChessMovesListViewRow.Height = (GridLength)gridLengthConverter.ConvertFromString(ChessAppSettings.Default.ChessMovesListViewRowHeight);
-
-            //if (!string.IsNullOrWhiteSpace(ChessAppSettings.Default.ChessMovesNotesRowHeight))
-            //    ChessMovesNotesRow.Height = (GridLength)gridLengthConverter.ConvertFromString(ChessAppSettings.Default.ChessMovesNotesRowHeight);
-
-            //PlayRadioButton.IsChecked = true;
-
-            //_chessMovesNotesRowHeight = ChessMovesNotesRow.Height;
-            //ChessMovesNotesRow.Height = new GridLength(0);
-            //HorizontalSplitterRow.Height = new GridLength(0);
         }
 
         /// <summary>
@@ -175,62 +161,11 @@ namespace Chess.View.Window
             ChessAppSettings.Default.Save();
         }
 
-        private void HorizontalGridSplitter_DragCompleted(object sender, System.Windows.Controls.Primitives.DragCompletedEventArgs e)
-        {
-            //var chessMovesListRowHeight = gridLengthConverter.ConvertToString(ChessMovesListViewRow.Height);
-            //var chessMovesNotesRowHeight = gridLengthConverter.ConvertToString(ChessMovesNotesRow.Height);
-
-            //ChessAppSettings.Default.ChessMovesListViewRowHeight = chessMovesListRowHeight;
-            //ChessAppSettings.Default.ChessMovesNotesRowHeight = chessMovesNotesRowHeight;
-            ChessAppSettings.Default.Save();
-        }
-
-        // private GridLength _chessMovesNotesRowHeight;
-
         private void PlayReCordReviewRadioButton_Click(object sender, RoutedEventArgs e)
         {
             var radioButton = sender as RadioButton;
-            var mode = radioButton?.Tag as string; 
-
-            //switch (mode)
-            //{
-            //    case "Play":
-            //        {
-            //            _chessMovesNotesRowHeight = ChessMovesNotesRow.Height;
-            //            ChessMovesNotesRow.Height = new GridLength(0);
-            //            HorizontalSplitterRow.Height = new GridLength(0);
-            //        }
-            //        break;
-            //    case "Record":
-            //        {
-            //            ChessMovesNotesRow.Height = _chessMovesNotesRowHeight;
-            //            HorizontalSplitterRow.Height = new GridLength(5);
-            //        }
-            //        break;
-            //    case "Review":
-            //        {
-            //            ChessMovesNotesRow.Height = _chessMovesNotesRowHeight;
-            //            HorizontalSplitterRow.Height = new GridLength(5);
-            //        }
-            //        break;
-            //    default:
-            //        throw new InvalidOperationException($"Unknown mode: {mode}");
-            //}
+            var mode = radioButton?.Tag as string;
         }
-
-        //// In MainWindow.xaml.cs, override OnPreviewMouseDown or OnMouseDown to handle routed MouseDown events:
-        //protected override void OnPreviewMouseDown(System.Windows.Input.MouseButtonEventArgs e)
-        //{
-        //    base.OnPreviewMouseDown(e);
-
-        //    // Check if the event originated from ChessMovesListViewBorder
-        //    var sourceBorder = e.OriginalSource as System.Windows.Controls.Border;
-        //    if (sourceBorder != null && sourceBorder.Name == "ChessMovesListViewBorder")
-        //    {
-        //        // Handle the event here
-        //        // Example: MessageBox.Show("MouseDown from ChessMovesListViewBorder handled in MainWindow!");
-        //    }
-        //}
 
         protected override void OnMouseDown(MouseButtonEventArgs e)
         {
@@ -238,13 +173,9 @@ namespace Chess.View.Window
 
             // Check if the event originated from ChessMovesListViewBorder
             var sourceBorder = e.OriginalSource as Border;
-            if (sourceBorder != null 
-                // && sourceBorder.Name == "ChessMovesListViewBorder"
-                )
+            if (sourceBorder != null)
             {
                 MainBoardGrid.Focus();
-                // Handle the event here
-                // Example: MessageBox.Show("MouseDown from ChessMovesListViewBorder handled in MainWindow!");
             }
         }
     }
